requirements.txt
discord.py
requests
flask
psutil
google-api-python-client
